requirements.txt
fastapi
sqlalchemy
python-dotenv
passlib
passlib[bcrypt]
python-jose[cryptography]
pydantic
pydantic[email]
uvicorn
pytest
pytest-asyncio
pytest-cov
bcrypt==4.0.1
slowapi
reportlab
httpx
requests
uuid
httpx
reportlab